requirements.txt
beautifulsoup4==4.11.2
blinker==1.8.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
requests==2.28.2
soupsieve==2.4
urllib3==1.26.14
Werkzeug==3.0.4
